feat(origin-ui-core): dont require refresh after metamask account change
