+++ packages/origin-ui-core/src/components/Modal/index.ts
@@ -1,3 +1,2 @@
-export * from './AccountChangedModal';
 export * from './RequestPasswordModal';
 export * from './RequestCertificatesModal';
