+++ packages/origin-ui-core/src/components/Table/PaginatedLoaderHooks.tsx
@@ -1,7 +1,46 @@
-import { ICustomFilter } from './FiltersHeader';
-import { PAGINATED_LOADER_INITIAL_STATE, DEFAULT_PAGE_SIZE } from './PaginatedLoader';
+import {
+    ICustomFilter,
+    RecordPropertyGetterFunction,
+    ITableColumn,
+    CustomFilterInputType
+} from '.';
 import { useState } from 'react';
-import { CurrentSortType } from './PaginatedLoaderFilteredSorted';
+import { IDeviceTypeService } from '@energyweb/utils-general';
+import { moment, Moment } from '../../utils';
+
+export const DEFAULT_PAGE_SIZE = 25;
+
+export interface IPaginatedLoaderFetchDataReturnValues {
+    paginatedData: any[];
+    total: number;
+}
+
+export interface IPaginatedLoaderState {
+    paginatedData: any[];
+    pageSize: number;
+    total: number;
+}
+
+export const PAGINATED_LOADER_INITIAL_STATE: IPaginatedLoaderState = {
+    paginatedData: [],
+    pageSize: DEFAULT_PAGE_SIZE,
+    total: 0
+};
+
+export type SortPropertiesType = ReadonlyArray<
+    | RecordPropertyGetterFunction
+    | readonly [
+          RecordPropertyGetterFunction,
+          (
+              value: ReturnType<RecordPropertyGetterFunction>
+          ) => ReturnType<RecordPropertyGetterFunction>
+      ]
+>;
+
+export type CurrentSortType = {
+    id: ITableColumn['id'];
+    sortProperties: ITableColumn['sortProperties'];
+};
 
 export interface IPaginatedLoaderHooksFetchDataParameters {
     requestedPageSize: number;
@@ -155,4 +194,91 @@ export function usePaginatedLoaderSorting<T>({
         currentSort,
         sortAscending
     };
+}
+
+export function checkRecordPassesFilters(
+    record: any,
+    filters: ICustomFilter[],
+    deviceTypeService?: IDeviceTypeService
+): boolean {
+    if (!filters) {
+        return true;
+    }
+
+    for (const filter of filters) {
+        const filteredPropertyResolvedValue = filter.property(record);
+
+        if (typeof filteredPropertyResolvedValue !== 'undefined') {
+            switch (filter.input.type) {
+                case CustomFilterInputType.string:
+                    if (
+                        filter.selectedValue &&
+                        !filteredPropertyResolvedValue
+                            .toString()
+                            .toLowerCase()
+                            .includes(filter.selectedValue.toLowerCase())
+                    ) {
+                        return false;
+                    }
+                    break;
+                case CustomFilterInputType.multiselect:
+                    return filter.selectedValue.includes(filteredPropertyResolvedValue);
+                case CustomFilterInputType.deviceType:
+                    if (!deviceTypeService) {
+                        throw new Error(
+                            `PaginatedLoaderFiltered requires "deviceTypeService" to be set to use "deviceType" filter`
+                        );
+                    }
+                    if (
+                        filter.selectedValue &&
+                        filter.selectedValue.length !== 0 &&
+                        !deviceTypeService.includesDeviceType(
+                            filteredPropertyResolvedValue?.toString(),
+                            filter.selectedValue as string[]
+                        )
+                    ) {
+                        return false;
+                    }
+                    break;
+                case CustomFilterInputType.dropdown:
+                    if (
+                        filter.selectedValue &&
+                        filter.selectedValue.toString() !== filteredPropertyResolvedValue.toString()
+                    ) {
+                        return false;
+                    }
+                    break;
+                case CustomFilterInputType.slider:
+                    if (filter.selectedValue) {
+                        const [min, max] = filter.selectedValue as number[];
+
+                        const valueAsNumber =
+                            typeof filteredPropertyResolvedValue === 'number'
+                                ? filteredPropertyResolvedValue
+                                : parseInt(filteredPropertyResolvedValue.toString(), 10);
+
+                        if (valueAsNumber < min || valueAsNumber > max) {
+                            return false;
+                        }
+                    }
+                    break;
+                case CustomFilterInputType.yearMonth:
+                    if (filter.selectedValue) {
+                        const year = (filter.selectedValue as Moment).year();
+                        const month = (filter.selectedValue as Moment).month();
+
+                        const recordDate = moment.unix(
+                            parseInt(filteredPropertyResolvedValue?.toString(), 10)
+                        );
+
+                        if (recordDate.month() !== month || recordDate.year() !== year) {
+                            return false;
+                        }
+                    }
+                    break;
+            }
+        }
+    }
+
+    return true;
 }
