+++ packages/origin-ui-core/src/components/Table/index.ts
@@ -1,3 +1,6 @@
 export * from './PaginatedLoaderHooks';
 export * from './TableMaterial';
-export { checkRecordPassesFilters } from './PaginatedLoaderFiltered';
+export * from './Actions';
+export * from './ColumnBatchActions';
+export * from './FiltersHeader';
+export * from './IndividualFilter';
