--- packages/origin-ui-core/src/features/contracts/selectors.ts
@@ -1,5 +1,0 @@
-import { IStoreState } from '../../types';
-import { IContractsLookup } from '@energyweb/origin-backend-core';
-
-export const getContractsLookup = (state: IStoreState): IContractsLookup =>
-    state.contracts.contractsLookup;
