+++ packages/origin-ui-core/src/features/general/sagas.ts
@@ -1,76 +1,69 @@
-import { call, put, delay, select, take, all, fork, cancelled, apply } from 'redux-saga/effects';
-import { Configuration } from '@energyweb/utils-general';
-import { SagaIterator } from 'redux-saga';
+import { call, put, select, take, all, fork, cancelled, apply, cancel } from 'redux-saga/effects';
+import { SagaIterator, eventChannel, EventChannel } from 'redux-saga';
 import {
-    hideAccountChangedModal,
-    showAccountChangedModal,
     setEnvironment,
     IEnvironment,
     GeneralActions,
     setOffChainDataSource,
     setExchangeClient,
-    setOffchainConfiguration
+    setOffchainConfiguration,
+    setError,
+    setLoading
 } from './actions';
-import { getConfiguration } from '../selectors';
-import {
-    getAccountChangedModalVisible,
-    getAccountChangedModalEnabled,
-    getEnvironment,
-    getOffChainDataSource
-} from './selectors';
-import { UsersActions } from '../users/actions';
+import { getEnvironment, getOffChainDataSource, getOffchainConfiguration } from './selectors';
 import axios, { Canceler } from 'axios';
 import { IOffChainDataSource, OffChainDataSource } from '@energyweb/origin-backend-client';
 import { ExchangeClient } from '../../utils/exchange';
-import { IOriginConfiguration } from '@energyweb/origin-backend-core';
+import {
+    IOriginConfiguration,
+    IOrganizationWithRelationsIds
+} from '@energyweb/origin-backend-core';
+import {
+    setActiveBlockchainAccountAddress,
+    addOrganizations,
+    UsersActions,
+    ISetActiveBlockchainAccountAddressAction
+} from '../users/actions';
+import { ethers } from 'ethers';
+import { getSearch } from 'connected-react-router';
+import { getConfiguration } from '../selectors';
+import * as queryString from 'query-string';
+import * as Winston from 'winston';
+import { Certificate, Contracts, CertificateUtils } from '@energyweb/issuer';
+import { Configuration, DeviceTypeService } from '@energyweb/utils-general';
+import { configurationUpdated } from '../actions';
+import { ProducingDevice } from '@energyweb/device-registry';
+import { producingDeviceCreatedOrUpdated } from '../producingDevices/actions';
+import { addCertificate, requestCertificateEntityFetch } from '../certificates/actions';
+import { IStoreState } from '../../types';
+import { BigNumber } from 'ethers/utils';
+
+function createEthereumProviderAccountsChangedEventChannel(ethereumProvider: any) {
+    return eventChannel<string[]>((emitter) => {
+        ethereumProvider.on('accountsChanged', emitter);
+
+        return () => {
+            ethereumProvider.off('accountsChanged', emitter);
+        };
+    });
+}
 
 function* showAccountChangedModalOnChange(): SagaIterator {
+    const ethereumProvider = (window as any)?.ethereum;
+
+    if (!ethereumProvider) {
+        return;
+    }
+
+    const channel: EventChannel<string[]> = yield call(
+        createEthereumProviderAccountsChangedEventChannel,
+        ethereumProvider
+    );
+
     while (true) {
-        yield take(UsersActions.setActiveBlockchainAccountAddress);
-        const conf: Configuration.Entity = yield select(getConfiguration);
+        const newAccounts: string[] = yield take(channel);
 
-        if (!conf) {
-            return;
-        }
-
-        try {
-            const initialAccounts: string[] = yield apply(
-                conf.blockchainProperties.web3,
-                conf.blockchainProperties.web3.listAccounts,
-                []
-            );
-
-            while (true) {
-                const accountChangedModalEnabled: boolean = yield select(
-                    getAccountChangedModalEnabled
-                );
-
-                if (!accountChangedModalEnabled) {
-                    break;
-                }
-
-                const accountChangedModalVisible: boolean = yield select(
-                    getAccountChangedModalVisible
-                );
-                const accounts: string[] = yield apply(
-                    conf.blockchainProperties.web3,
-                    conf.blockchainProperties.web3.listAccounts,
-                    []
-                );
-
-                if (accountChangedModalVisible) {
-                    if (initialAccounts[0] === accounts[0]) {
-                        yield put(hideAccountChangedModal());
-                    }
-                } else if (initialAccounts[0] !== accounts[0]) {
-                    yield put(showAccountChangedModal());
-                }
-
-                yield delay(1000);
-            }
-        } catch (error) {
-            console.error('showAccountChangedModalOnChange() error', error);
-        }
+        yield put(setActiveBlockchainAccountAddress((newAccounts && newAccounts[0]) || ''));
     }
 }
 
@@ -182,11 +175,229 @@ function* initializeOffChainDataSource(): SagaIterator {
     }
 }
 
+enum ERROR {
+    WRONG_NETWORK_OR_CONTRACT_ADDRESS = "Please make sure you've chosen correct blockchain network and the contract address is valid."
+}
+
+function createConfiguration(
+    account: string,
+    web3: ethers.providers.JsonRpcProvider,
+    offchainConfiguration: IOriginConfiguration,
+    offChainDataSource: IOffChainDataSource,
+    logger: Winston.Logger = Winston.createLogger({
+        level: 'verbose',
+        format: Winston.format.combine(Winston.format.colorize(), Winston.format.simple()),
+        transports: [new Winston.transports.Console({ level: 'silly' })]
+    })
+) {
+    const signer = web3.getSigner(account);
+    const { contractsLookup } = offchainConfiguration;
+
+    const blockchainProperties: Configuration.BlockchainProperties = {
+        registry: Contracts.factories.RegistryFactory.connect(contractsLookup.registry, signer),
+        issuer: Contracts.factories.IssuerFactory.connect(contractsLookup.issuer, signer),
+        web3,
+        activeUser: signer
+    };
+
+    return {
+        blockchainProperties,
+        offChainDataSource,
+        logger,
+        deviceTypeService: new DeviceTypeService(offchainConfiguration.deviceTypes)
+    };
+}
+
+async function initConf(
+    routerSearch: string,
+    offchainConfiguration: IOriginConfiguration,
+    offChainDataSource: IOffChainDataSource,
+    environmentWeb3: string
+): Promise<IStoreState['configuration']> {
+    let web3: ethers.providers.JsonRpcProvider = null;
+    const params = queryString.parse(routerSearch);
+
+    const ethereumProvider = (window as any).ethereum;
+
+    if (params.rpc) {
+        web3 = new ethers.providers.JsonRpcProvider(params.rpc as string);
+    } else if (ethereumProvider) {
+        web3 = new ethers.providers.Web3Provider(ethereumProvider);
+        try {
+            // Request account access if needed
+            await ethereumProvider.enable();
+        } catch (error) {
+            // User denied account access...
+        }
+    } else if ((window as any).web3) {
+        web3 = new ethers.providers.Web3Provider((window as any).web3.currentProvider);
+    } else if (environmentWeb3) {
+        web3 = new ethers.providers.JsonRpcProvider(environmentWeb3);
+    }
+
+    const accounts = await web3.listAccounts();
+
+    return createConfiguration(accounts[0], web3, offchainConfiguration, offChainDataSource);
+}
+
+function* initEventHandler() {
+    const configuration: IStoreState['configuration'] = yield select(getConfiguration);
+
+    if (!configuration) {
+        return;
+    }
+
+    configuration.blockchainProperties.registry
+        .on(
+            'ClaimSingle',
+            async (
+                _claimIssuer: string,
+                _claimSubject: string,
+                _topic: BigNumber,
+                _id: BigNumber
+            ) => {
+                put(requestCertificateEntityFetch(_id.toNumber()));
+            }
+        )
+        .on('IssuanceSingle', async (sender: string, topic: BigNumber, id: BigNumber) => {
+            put(requestCertificateEntityFetch(id?.toNumber()));
+        });
+}
+
+function* fillContractLookupIfMissing(): SagaIterator {
+    while (true) {
+        yield take([
+            GeneralActions.setEnvironment,
+            GeneralActions.setOffChainDataSource,
+            GeneralActions.setOffchainConfiguration
+        ]);
+
+        const environment: IEnvironment = yield select(getEnvironment);
+        const offChainDataSource: IOffChainDataSource = yield select(getOffChainDataSource);
+        const offchainConfiguration: IOriginConfiguration = yield select(getOffchainConfiguration);
+
+        if (!environment || !offChainDataSource || !offchainConfiguration) {
+            continue;
+        }
+
+        if (!offchainConfiguration.contractsLookup) {
+            yield put(setError(ERROR.WRONG_NETWORK_OR_CONTRACT_ADDRESS));
+            yield put(setLoading(false));
+
+            continue;
+        }
+
+        const routerSearch: string = yield select(getSearch);
+
+        let configuration: IStoreState['configuration'];
+        try {
+            configuration = yield call(
+                initConf,
+                routerSearch,
+                offchainConfiguration,
+                offChainDataSource,
+                environment.WEB3
+            );
+
+            const userAddress = yield apply(
+                configuration.blockchainProperties.activeUser,
+                configuration.blockchainProperties.activeUser.getAddress,
+                []
+            );
+
+            yield put(configurationUpdated(configuration));
+            yield put(setActiveBlockchainAccountAddress(userAddress));
+
+            yield put(setLoading(false));
+        } catch (error) {
+            console.error('ContractsSaga::WrongNetwork', error);
+            yield put(setError(ERROR.WRONG_NETWORK_OR_CONTRACT_ADDRESS));
+            yield put(setLoading(false));
+
+            yield cancel();
+        }
+
+        try {
+            const producingDevices: ProducingDevice.Entity[] = yield apply(
+                ProducingDevice,
+                ProducingDevice.getAllDevices,
+                [configuration]
+            );
+
+            for (const device of producingDevices) {
+                yield put(producingDeviceCreatedOrUpdated(device));
+            }
+
+            const certificates: Certificate[] = yield apply(
+                Certificate,
+                CertificateUtils.getAllCertificates,
+                [configuration]
+            );
+
+            for (const certificate of certificates) {
+                yield put(addCertificate(certificate));
+            }
+
+            const organizations: IOrganizationWithRelationsIds[] = yield apply(
+                offChainDataSource.organizationClient,
+                offChainDataSource.organizationClient.getAll,
+                []
+            );
+
+            yield put(addOrganizations(organizations));
+
+            yield call(initEventHandler);
+        } catch (error) {
+            console.error('fillContractLookupIfMissing() error', error);
+        }
+    }
+}
+
+function* updateConfigurationWhenUserChanged(): SagaIterator {
+    while (true) {
+        const { payload: address }: ISetActiveBlockchainAccountAddressAction = yield take(
+            UsersActions.setActiveBlockchainAccountAddress
+        );
+
+        const existingConfiguration: Configuration.Entity = yield select(getConfiguration);
+        const offchainConfiguration: IOriginConfiguration = yield select(getOffchainConfiguration);
+
+        const existingSignerAddress: string = yield call([
+            existingConfiguration?.blockchainProperties?.activeUser,
+            existingConfiguration?.blockchainProperties?.activeUser?.getAddress
+        ]);
+
+        if (
+            !existingConfiguration ||
+            !offchainConfiguration ||
+            address?.toLowerCase() === existingSignerAddress?.toLowerCase()
+        ) {
+            continue;
+        }
+
+        const web3 = existingConfiguration?.blockchainProperties?.web3;
+        const offChainDataSource = existingConfiguration?.offChainDataSource;
+
+        const newConfiguration: Configuration.Entity = yield call(
+            createConfiguration,
+            address,
+            web3,
+            offchainConfiguration,
+            offChainDataSource,
+            existingConfiguration.logger
+        );
+
+        yield put(configurationUpdated(newConfiguration));
+    }
+}
+
 export function* generalSaga(): SagaIterator {
     yield all([
         fork(showAccountChangedModalOnChange),
         fork(setupEnvironment),
         fork(initializeOffChainDataSource),
-        fork(fillOffchainConfiguration)
+        fork(fillOffchainConfiguration),
+        fork(fillContractLookupIfMissing),
+        fork(updateConfigurationWhenUserChanged)
     ]);
 }
