+++ packages/origin-ui-core/src/features/general/selectors.ts
@@ -1,12 +1,6 @@
 import { IStoreState } from '../../types';
 import { ExternalDeviceIdType } from '@energyweb/origin-backend-core';
 
-export const getAccountChangedModalVisible = (state: IStoreState) =>
-    state.general.accountChangedModalVisible;
-
-export const getAccountChangedModalEnabled = (state: IStoreState) =>
-    state.general.accountChangedModalEnabled;
-
 export const getLoading = (state: IStoreState) => state.general.loading;
 
 export const getError = (state: IStoreState) => state.general.error;
