+++ packages/origin-ui-core/src/features/sagas.ts
@@ -1,10 +1,8 @@
 import { generalSaga } from './general/sagas';
-import { contractsSaga } from './contracts/sagas';
 import { usersSaga } from './users/sagas';
 import { certificatesSaga } from './certificates/sagas';
 
 export const sagas = {
-    contractsSaga,
     generalSaga,
     usersSaga,
     certificatesSaga
