+++ packages/origin-ui-core/src/reducers/index.tsx
@@ -2,7 +2,6 @@ import certificates from '../features/certificates/reducer';
 import producingDevices from '../features/producingDevices/reducer';
 import configuration from './Configuration';
 import general from '../features/general/reducer';
-import contracts from '../features/contracts/reducer';
 import users from '../features/users/reducer';
 import { combineReducers } from 'redux';
 import { IStoreState } from '../types';
@@ -14,7 +13,6 @@ export const createRootReducer = (history) =>
         producingDevices,
         general,
         configuration,
-        contracts,
         users,
         router: connectRouter(history)
     });
