+++ packages/origin-ui-core/src/types/index.tsx
@@ -1,6 +1,5 @@
 import { Configuration } from '@energyweb/utils-general';
 import { IGeneralState } from '../features/general/reducer';
-import { IContractsState } from '../features/contracts/reducer';
 import { ICertificatesState } from '../features/certificates/reducer';
 import { IUsersState } from '../features/users/reducer';
 import { IProducingDevicesState } from '../features/producingDevices/reducer';
@@ -11,7 +10,6 @@ export interface IStoreState {
     producingDevices: IProducingDevicesState;
     certificates: ICertificatesState;
     general: IGeneralState;
-    contracts: IContractsState;
     users: IUsersState;
     router: RouterState;
 }
